utils/requirements.txt
numpy==1.17.2
pillow==6.2.0
pytest==5.2.1
pytest-cov==2.8.1
tqdm==4.36.1